requirements.txt
BareNecessities==0.2.8
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
config==0.5.1
dominate==2.8.0
Flask==2.3.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
google==3.0.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleAPI==0.1.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mail==2.1.0
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
protobuf==4.25.1
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.16
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.2
visitor==0.1.3
Werkzeug==2.3.5
WTForms==3.0.1
